fix(ui): resolve 7 UI/UX issues across mobileapp
- _layout.tsx: load Outfit_600SemiBold (was missing, caused silent
  font fallback on all buttons, inputs, labels across every screen)

- settings.tsx: fix SafeAreaView import (react-native -> safe-area-context),
  load real username + wallet address from storage/keypair,
  persist biometrics toggle to AsyncStorage (was resetting on every visit)

- receive.tsx: replace deprecated Clipboard from react-native with
  expo-clipboard, load real wallet address from local keypair instead
  of hardcoded placeholder GABC1234EXAMPLE...

- AccountTypeCard.tsx: borderRadius 100 -> 16 (pill on a tall card
  looked wrong and clipped icon container)

- YieldHomeCard.tsx: replace 3 hardcoded hex colors (#0F3D2E,
  #34D399, #9FD9B5, #BFE9CF) with COLORS constants

- colors.ts: add yieldBg, yieldAccent, yieldLabel, yieldStatus
  constants for the earnings card palette

- backend/src/main.rs: fix health probe SELECT 1 type (i64 -> i32,
  Postgres returns INT4 not INT8 for literal 1)

+++ backend/src/main.rs
@@ -318,12 +318,11 @@ async fn health_check(State(state): State<HealthState>) -> impl IntoResponse {
 /// Basic Postgres connectivity: a single round-trip to the DB pool.
 async fn probe_database(pool: &sqlx::PgPool) -> DbHealth {
     let start = Instant::now();
-    match sqlx::query_scalar::<_, i64>("SELECT 1")
+    match sqlx::query_scalar::<_, i32>("SELECT 1")
         .fetch_one(pool)
         .await
     {
-        Ok(_) => DbHealth {
-            status: "ok",
+        Ok(_) => DbHealth {            status: "ok",
             latency_ms: start.elapsed().as_millis() as u64,
             error: None,
         },
